Use Berkeley DB 4.8
Using Berkeley DB 4.8 should ensure wallet compatibility across other wallets. Unless the Windows/Mac wallets are compiled with BerkeleyDB 5 or higher, in which case ¯\_(ツ)_/¯.

In any case, this is the recommended build style for wallets. And if you happened to use any of the AppImage test builds, those wallets will be incompatible with this build and you will have to send any coins in those wallets to a new one.